Add gas entities
